sync the latest code

+++ GDSConHand/src/com/winjune/wifiindoor/navi/NaviData.java
@@ -26,7 +26,7 @@ public class NaviData implements Serializable {
 		this.to = to;
 	}
 
-	public float getDistance() {
+	public int getDistance() {
 		return distance;
 	}
